refactor get_abbreviated_xpath to avoid circular imports
