+++ onadata/libs/serializers/merged_xform_serializer.py
@@ -15,8 +15,9 @@ from pyxform.errors import PyXFormError
 from rest_framework import serializers
 
 from onadata.apps.logger.models import MergedXForm, XForm
-from onadata.apps.logger.models.xform import XFORM_TITLE_LENGTH, get_abbreviated_xpath
+from onadata.apps.logger.models.xform import XFORM_TITLE_LENGTH
 from onadata.libs.utils.common_tags import MULTIPLE_SELECT_TYPE, SELECT_ONE
+from onadata.libs.utils.common_tools import get_abbreviated_xpath
 
 SELECTS = [SELECT_ONE, MULTIPLE_SELECT_TYPE]
